Work App.

+++ src/main/java/ru/sladkkov/database/springEVM/repository/BuyerRepository.java
@@ -3,5 +3,5 @@ package ru.sladkkov.database.springEVM.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.sladkkov.database.springEVM.model.Buyer;
 
-public interface BuyerRepository extends JpaRepository<Buyer, Long> {
+public interface BuyerRepository extends JpaRepository<Buyer, Integer> {
 }
